Added memory leak check in subcc classes. Adjusted empty lines between test result output of each test routine.

+++ cpp/lib_calvin/subcc/syntax_tree.h
@@ -46,7 +46,7 @@ class SyntaxTreeNode {
   protected:
     SyntaxTreeNode const *parent_;
 	private:
-		static size_t objectCount_;
+		static int64_t objectCount_;
 };
 
 // Special nodes for storing dimension of an array declaration
